+++ crates/cli/tests/utils/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2022 Shun Sakai
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+pub mod command;
